+++ packages/f-checkout/src/components/Selector.vue
@@ -1,7 +1,7 @@
 <template>
     <form-dropdown
         id="time-selection"
-        data-test-id="form-select"
+        name="order-time"
         :class="$style['c-checkout-selector']"
         input-type="dropdown"
         :label-text="orderMethod"